refactor: replace TextScope with StringBuilder context extensions
Drop the buildText/TextScope wrapper in favor of context-parameter
extensions on StringBuilder, so the standard buildString can be used
directly with the unaryPlus and trimLastNewLine helpers.
